refactor: :truck: footprints 페이지 app 라우터로 옮김
