+++ src/components/index.ts
@@ -3,7 +3,10 @@ export { Layout } from './Layout';
 export { LinkButton } from './LinkButton';
 export { BadgeList } from './BadgeList';
 export { Paragraph } from './Paragraph';
+export { DescriptionList } from './DescriptionList';
 export { BranchListWrapper } from './BranchListWrapper';
 
 export { ProjectElement } from './ProjectElement';
 export { ProjectComponent } from './ProjectComponent';
+
+export { FootprintElement } from './FootprintElement';
